django/vicky/requirements.txt
Django==3.2.*
requests==2.26.0
django-tinymce==4.0.0
Pillow==9.0.
django-autoslug==1.9.9
django-bootstrap-v5==1.0.11
psycopg2==2.9.9
django-tinymce==4.0.0



